IGNITE-28938 Separate walking a message from marshalling it
The generated companion did both: it turned @Marshalled fields into bytes and walked the nested messages and cache
objects. Only 28 of the 197 it made for core marshalled anything, so the name fitted almost none of them.

The two are companions of their own now. MessageMarshaller keeps the bytes: @Marshalled fields and the marshal call
a MarshallableMessage defines. MessageWireForm walks the fields. MessageWireGenerator and MessageMarshallerGenerator
generate them over the shared ground of MessageWireCompanionGenerator, and MessageWire runs the three steps in order:
the message's own step, marshalling, walking on the way out, and the reverse on the way in.

Core now gets 28 marshallers instead of 197, and 187 wire forms.
