+++ modules/codegen/src/main/java/org/apache/ignite/internal/MessageMarshallerGenerator.java
@@ -21,43 +21,29 @@ import java.io.IOException;
 import java.io.StringWriter;
 import java.io.Writer;
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.function.BiFunction;
 import javax.annotation.processing.ProcessingEnvironment;
 import javax.lang.model.element.Element;
 import javax.lang.model.element.Modifier;
 import javax.lang.model.element.QualifiedNameable;
-import javax.lang.model.element.TypeElement;
 import javax.lang.model.element.VariableElement;
 import javax.lang.model.type.ArrayType;
 import javax.lang.model.type.DeclaredType;
 import javax.lang.model.type.TypeKind;
 import javax.lang.model.type.TypeMirror;
-import javax.lang.model.type.TypeVariable;
-import javax.lang.model.util.ElementFilter;
-import javax.tools.Diagnostic;
-import org.apache.ignite.internal.systemview.SystemViewRowAttributeWalkerProcessor;
-import org.jetbrains.annotations.Nullable;
 
-import static org.apache.ignite.internal.MessageProcessor.CACHE_OBJECT_CLS;
 import static org.apache.ignite.internal.MessageProcessor.IGNITE_CHECKED_EXCEPTION_CLS;
 import static org.apache.ignite.internal.MessageProcessor.KEY_CACHE_OBJECT_CLS;
 import static org.apache.ignite.internal.MessageProcessor.MARSHALLABLE_MESSAGE_INTERFACE;
-import static org.apache.ignite.internal.MessageProcessor.MESSAGE_INTERFACE;
-import static org.apache.ignite.internal.MessageProcessor.NON_MARSHALLABLE_MESSAGE_INTERFACE;
 
 /**
- * Generates the {@code *Marshaller} class of a {@code Message}: the code that marshals its fields, walks into its
- * nested messages, and prepares its cache objects. A message with none of that gets no marshaller. A step the message
- * defines itself, such as {@code CustomWireFormMessage}, is not marshalling and is run by {@code MessageWire} instead.
+ * Generates the {@code *Marshaller} class of a {@code Message}: the code that turns a {@code @Marshalled} field into
+ * bytes and back, and calls the {@code marshal} a {@code MarshallableMessage} defines. Walking the fields is a step of
+ * its own and belongs to {@link MessageWireFormGenerator}; a message with nothing to marshal gets no marshaller.
  */
-public class MessageMarshallerGenerator extends MessageCompanionGenerator {
+public class MessageMarshallerGenerator extends MessageWireCompanionGenerator {
     /** Interface the generated marshallers implement. */
     private static final String MESSAGE_MARSHALLER_CLS = "org.apache.ignite.plugin.extensions.communication.MessageMarshaller";
 
@@ -65,79 +51,19 @@ public class MessageMarshallerGenerator extends MessageCompanionGenerator {
     private static final String MARSHALLER_CLS = "org.apache.ignite.marshaller.Marshaller";
 
     /** */
-    private static final String GRID_KERNAL_CONTEXT_CLS = "org.apache.ignite.internal.GridKernalContext";
-
-    /** */
-    private static final String CACHE_OBJECT_CONTEXT_CLS = "org.apache.ignite.internal.processors.cache.CacheObjectContext";
-
-    /** */
-    private static final String GRID_CACHE_GROUP_ID_MESSAGE_CLS = "org.apache.ignite.internal.processors.cache.GridCacheGroupIdMessage";
-
-    /** Facade the generated code calls to take nested messages to the wire and back. */
-    private static final String MESSAGE_WIRE_CLS = "org.apache.ignite.internal.managers.communication.MessageWire";
-
-    /** */
-    private static final String IGNITE_MESSAGE_FACTORY_CLS = "org.apache.ignite.internal.managers.communication.IgniteMessageFactory";
-
-    /** {@code IgniteUtils} shortcut used by the generated {@code @Marshalled} handling. */
-    private static final String U_CLS = "org.apache.ignite.internal.util.typedef.internal.U";
-
-    /** Accumulated source lines for all generated marshal/unmarshal methods. */
-    private final List<String> marshall = new ArrayList<>();
-
-    /** */
     private final TypeMirror marshallableMsgType;
 
-    /** */
-    private final TypeMirror msgType;
-
-    /** */
-    private final TypeMirror cacheObjType;
-
-    /** */
-    private final TypeMirror nonMarshallableType;
-
-    /** */
-    private final TypeMirror cacheGrpIdMsgType;
-
-    /** */
-    private final TypeMirror mapType;
-
-    /** */
-    private final TypeMirror colType;
-
-    /** */
+    /** Whether the message marshals some of its fields itself. */
     private boolean marshallable;
 
-    /** */
+    /** Whether the message has a {@code @Marshalled} field that becomes bytes. */
     private boolean hasMarshalled;
-
-    /** Whether any generated method got a non-empty body; a marshaller without one is skipped entirely. */
-    private boolean hasStatements;
-
-    /** Enclosed fields of the currently processed type. Computed once per {@link #generateBody} call. */
-    private Map<String, VariableElement> enclosed;
-
-    /** {@link MarshalledKind} of each {@code @Marshalled} enclosed field. Computed once per {@link #generateBody} call. */
-    private final Map<VariableElement, MarshalledKind> kinds = new HashMap<>();
-
-    /** Nesting depth of the current marshal for-loop; names loop variables {@code e}, {@code e1}, {@code e2}… */
-    private int loopDepth;
-
-    /** Whether the currently generated method emitted a loop-nested facade call and so needs the {@code msgFactory} local. */
-    private boolean usesMsgFactory;
 
     /** */
     MessageMarshallerGenerator(ProcessingEnvironment env) {
         super(env);
 
         marshallableMsgType = type(MARSHALLABLE_MESSAGE_INTERFACE);
-        msgType = type(MESSAGE_INTERFACE);
-        cacheObjType = type(CACHE_OBJECT_CLS);
-        nonMarshallableType = type(NON_MARSHALLABLE_MESSAGE_INTERFACE);
-        cacheGrpIdMsgType = type(GRID_CACHE_GROUP_ID_MESSAGE_CLS);
-        mapType = type(Map.class.getName());
-        colType = type(Collection.class.getName());
     }
 
     /** {@inheritDoc} */
@@ -146,30 +72,18 @@ public class MessageMarshallerGenerator extends MessageCompanionGenerator {
     }
 
     /** {@inheritDoc} */
-    @Override protected boolean shouldSkip(TypeElement type, List<VariableElement> fields) {
-        return isNonMarshallable(type.asType());
-    }
-
-    /** {@inheritDoc} */
     @Override protected void generateBody(List<VariableElement> fields) {
-        enclosed = enclosedFields();
-
-        for (VariableElement f : enclosed.values()) {
-            MarshalledKind kind = marshalledKind(f);
-
-            if (kind != null)
-                kinds.put(f, kind);
-        }
+        readFields();
 
         marshallable = marshallableMsgType != null && assignableFrom(type.asType(), marshallableMsgType);
         hasMarshalled = kinds.values().stream().anyMatch(k -> k == MarshalledKind.BLOB || k == MarshalledKind.ELEMENT_BLOBS);
 
         generateMarshalMethod(fields);
-        generateUnmarshalMethods(fields);
+        generateUnmarshalMethod(fields);
     }
 
     /** {@inheritDoc} */
-    @Override protected String buildClassCode(String marshallerClsName) throws IOException {
+    @Override protected String buildClassCode(String clsName) throws IOException {
         if (!hasStatements)
             return null;
 
@@ -180,19 +94,83 @@ public class MessageMarshallerGenerator extends MessageCompanionGenerator {
             if (marshallable || hasMarshalled)
                 imports.add(MARSHALLER_CLS);
 
-            writeClassHeader(writer, "MessageMarshaller", marshallerClsName);
+            writeClassHeader(writer, "MessageMarshaller", clsName);
 
             writer.write(" {" + NL);
 
-            writeConstructor(writer, marshallerClsName);
+            writeConstructor(writer, clsName);
 
-            for (String line : marshall)
+            for (String line : methods)
                 writer.write(line + NL);
 
             writer.write("}");
 
             return writer.toString();
         }
+    }
+
+    /** Generates the {@code marshal} method: the fields that become bytes, plus the message's own call. */
+    private void generateMarshalMethod(List<VariableElement> orderedFields) {
+        imports.add(IGNITE_CHECKED_EXCEPTION_CLS);
+        imports.add(GRID_KERNAL_CONTEXT_CLS);
+        imports.add(CACHE_OBJECT_CONTEXT_CLS);
+
+        if (!kinds.isEmpty())
+            imports.add(U_CLS);
+
+        String signature = "marshal(" + simpleNameWithGeneric(type) + " msg, GridKernalContext kctx, CacheObjectContext cacheObjCtx)";
+
+        hasStatements |= emitMethod(methods, signature, body -> {
+            usesMsgFactory = false;
+
+            List<String> code = new ArrayList<>();
+
+            appendMarshalledFieldsPrepare(code);
+            appendMarshalledPrepare(code);
+
+            if (marshallable)
+                appendBlock(code, List.of(indentedLine("msg.marshal(marshaller);")));
+
+            if (code.isEmpty())
+                return;
+
+            if (needsCtx(orderedFields))
+                appendBlock(body, List.of(ctxResolutionLine()));
+
+            body.addAll(code);
+
+            prependMsgFactoryResolution(body);
+        });
+    }
+
+    /** Generates the {@code unmarshal} method: the message's own call, then the fields rebuilt from bytes. */
+    private void generateUnmarshalMethod(List<VariableElement> orderedFields) {
+        String params = simpleNameWithGeneric(type) + " msg, GridKernalContext kctx, CacheObjectContext cacheObjCtx, ClassLoader clsLdr";
+
+        hasStatements |= emitMethod(methods, "unmarshal(" + params + ")", body -> {
+            usesMsgFactory = false;
+
+            List<String> code = new ArrayList<>();
+
+            if (marshallable)
+                appendBlock(code, List.of(indentedLine("msg.unmarshal(marshaller, clsLdr);")));
+
+            appendMarshalledFinish(code);
+
+            appendMarshalledElementsFinish(code);
+            appendMarshalledMapFinish(code);
+            appendMarshalledElementBlobsFinish(code);
+
+            if (code.isEmpty())
+                return;
+
+            if (needsCtx(orderedFields) || !marshalledWireFieldsToSkip().isEmpty())
+                appendBlock(body, List.of(ctxResolutionLine()));
+
+            body.addAll(code);
+
+            prependMsgFactoryResolution(body);
+        });
     }
 
     /** Writes the {@code marshaller} field and the constructor initializing it, when the marshaller is needed. */
@@ -219,121 +197,6 @@ public class MessageMarshallerGenerator extends MessageCompanionGenerator {
 
         writer.write(indentedLine("}"));
         writer.write(NL + NL);
-    }
-
-    /** Generates the {@code marshal} method body and appends it to {@link #marshall}. */
-    private void generateMarshalMethod(List<VariableElement> orderedFields) {
-        imports.add(IGNITE_CHECKED_EXCEPTION_CLS);
-        imports.add(GRID_KERNAL_CONTEXT_CLS);
-        imports.add(CACHE_OBJECT_CONTEXT_CLS);
-
-        String signature = "marshal(" + simpleNameWithGeneric(type) + " msg, GridKernalContext kctx, CacheObjectContext cacheObjCtx)";
-
-        hasStatements |= emitMethod(marshall, signature, body -> {
-            usesMsgFactory = false;
-
-            if (needsCtx(orderedFields))
-                appendBlock(body, List.of(ctxResolutionLine()));
-
-            appendMarshalledFieldsPrepare(body);
-            appendMarshalledPrepare(body);
-
-            if (marshallable)
-                appendBlock(body, List.of(indentedLine("msg.marshal(marshaller);")));
-
-            appendFields(body, orderedFields, Direction.OUT);
-
-            prependMsgFactoryResolution(body);
-        });
-    }
-
-    /** Generates all {@code unmarshal} overloads and appends them to {@link #marshall}. */
-    private void generateUnmarshalMethods(List<VariableElement> orderedFields) {
-        List<VariableElement> nioFields = new ArrayList<>();
-        List<VariableElement> workerFields = new ArrayList<>();
-
-        for (VariableElement f : orderedFields) {
-            boolean nioField = isNioField(f);
-
-            if (nioField && isMessage(f.asType()) && !nestedNeedsCtx(f.asType()))
-                nioFields.add(f);
-            else {
-                if (nioField && !isMessage(f.asType())) {
-                    env.getMessager().printMessage(Diagnostic.Kind.ERROR,
-                        "@NioField has no effect on non-Message field '" + f.getSimpleName() + "' of type " + f.asType(), f);
-                }
-                else if (nioField) {
-                    env.getMessager().printMessage(Diagnostic.Kind.ERROR,
-                        "@NioField field '" + f.getSimpleName() + "' of type " + f.asType() + " needs a cache object " +
-                            "context to unmarshal, but the NIO thread has none; only context-free messages may be @NioField", f);
-                }
-
-                workerFields.add(f);
-            }
-        }
-
-        // U is referenced only by @Marshalled handling.
-        boolean usesU = enclosed.values().stream().anyMatch(f -> f.getAnnotation(Marshalled.class) != null);
-
-        if (usesU)
-            imports.add(U_CLS);
-
-        String msgParam = simpleNameWithGeneric(type) + " msg, GridKernalContext kctx";
-
-        generateUnmarshalMethod(msgParam + ", CacheObjectContext cacheObjCtx, ClassLoader clsLdr", workerFields);
-
-        if (!nioFields.isEmpty())
-            generateUnmarshalNioMethod(msgParam, nioFields);
-    }
-
-    /** Generates the cache-aware {@code unmarshal} overload: the full field set, with cache context and deployment class loader. */
-    private void generateUnmarshalMethod(String params, List<VariableElement> fields) {
-        hasStatements |= emitMethod(marshall, "unmarshal(" + params + ")", body -> {
-            usesMsgFactory = false;
-
-            Set<String> wireFieldSkip = marshalledWireFieldsToSkip();
-
-            if (needsCtx(fields) || !wireFieldSkip.isEmpty())
-                appendBlock(body, List.of(ctxResolutionLine()));
-
-            appendFields(body, fields, Direction.IN, wireFieldSkip);
-
-            if (marshallable)
-                appendBlock(body, List.of(indentedLine("msg.unmarshal(marshaller, clsLdr);")));
-
-            appendMarshalledFinish(body);
-
-            appendMarshalledElementsFinish(body);
-            appendMarshalledMapFinish(body);
-            appendMarshalledElementBlobsFinish(body);
-
-            prependMsgFactoryResolution(body);
-        });
-    }
-
-    /** Generates the {@code unmarshalNio} method for NIO-eligible {@code @Message} fields. */
-    private void generateUnmarshalNioMethod(String params, List<VariableElement> nioFields) {
-        hasStatements |= emitMethod(marshall, "unmarshalNio(" + params + ")", body -> {
-            for (VariableElement f : nioFields)
-                appendBlock(body, fromWireNioField(fieldAccessor(f)));
-        });
-    }
-
-    /** Cache-free unmarshal of a {@code @NioField} message field on the NIO thread (no cache context available). */
-    private List<String> fromWireNioField(String accessor) {
-        imports.add(MESSAGE_WIRE_CLS);
-
-        List<String> code = new ArrayList<>();
-
-        code.add(indentedLine("if (%s != null)", accessor));
-
-        indent++;
-
-        code.add(indentedLine("MessageWire.fromWire(%s, kctx);", accessor));
-
-        indent--;
-
-        return code;
     }
 
     /** Generates logical→wire conversions for the element-, blob-list- and map-flavoured {@code @Marshalled} fields. */
@@ -589,29 +452,6 @@ public class MessageMarshallerGenerator extends MessageCompanionGenerator {
         return code;
     }
 
-    /** Returns names of wire fields skipped by {@link #appendFields} in UNMARSHAL mode. */
-    private Set<String> marshalledWireFieldsToSkip() {
-        Set<String> names = new HashSet<>();
-
-        for (VariableElement f : enclosed.values()) {
-            MarshalledKind kind = kinds.get(f);
-
-            if (kind == null || kind == MarshalledKind.BLOB)
-                continue;
-
-            Marshalled ann = f.getAnnotation(Marshalled.class);
-
-            if (kind == MarshalledKind.MAP) {
-                names.add(ann.keys());
-                names.add(ann.values());
-            }
-            else
-                names.add(ann.value());
-        }
-
-        return names;
-    }
-
     /** Generates Map reconstruction for all map-flavoured {@code @Marshalled} fields. */
     private void appendMarshalledMapFinish(List<String> body) {
         for (VariableElement field : enclosed.values()) {
@@ -788,419 +628,5 @@ public class MessageMarshallerGenerator extends MessageCompanionGenerator {
         inner.add(indentedLine("%s = %s.values();", valuesField, mapField));
 
         return inner;
-    }
-
-    /** Marshals each field and appends non-empty results to {@code body}. */
-    private void appendFields(List<String> body, List<VariableElement> fields, Direction mode) {
-        appendFields(body, fields, mode, Set.of());
-    }
-
-    /** Marshals each field, skipping names in {@code skip}, and appends non-empty results to {@code body}. */
-    private void appendFields(List<String> body, List<VariableElement> fields, Direction mode, Set<String> skip) {
-        for (VariableElement field : fields) {
-            if (skip.contains(field.getSimpleName().toString()))
-                continue;
-
-            List<String> result = codeFor(field.asType(), fieldAccessor(field), mode);
-
-            if (!result.isEmpty())
-                appendBlock(body, result);
-        }
-    }
-
-    /** Returns generated marshal/unmarshal code lines for field of type {@code t}, or empty if none needed. */
-    private List<String> codeFor(TypeMirror t, String accessor, Direction mode) {
-        if (t.getKind() == TypeKind.ARRAY) {
-            TypeMirror comp = ((ArrayType)t).getComponentType();
-
-            return comp.getKind() == TypeKind.DECLARED ? arrayCode(comp, accessor, mode) : List.of();
-        }
-
-        if (t.getKind() == TypeKind.DECLARED || t.getKind() == TypeKind.TYPEVAR) {
-            if (isMessage(t))
-                return isNonMarshallable(t) ? List.of() : messageCode(accessor, mode);
-            if (isCacheObject(t))
-                return cacheObjectCode(accessor, mode);
-            if (isMap(t))
-                return mapCode((DeclaredType)t, accessor, mode);
-            if (isCollection(t))
-                return collectionCode((DeclaredType)t, accessor, mode);
-        }
-
-        return List.of();
-    }
-
-    /**
-     * Generates a null-guarded {@code MessageWire} call. Loop-nested calls go through the overloads taking the
-     * pre-resolved {@code msgFactory} local (see {@link #prependMsgFactoryResolution}), so the factory is not
-     * re-resolved from the context on every element.
-     */
-    private List<String> messageCode(String accessor, Direction mode) {
-        imports.add(MESSAGE_WIRE_CLS);
-
-        List<String> code = new ArrayList<>();
-
-        code.add(indentedLine("if (%s != null)", accessor));
-
-        indent++;
-
-        if (loopDepth > 0) {
-            usesMsgFactory = true;
-
-            code.add(mode == Direction.OUT
-                ? indentedLine("MessageWire.toWire(msgFactory, %s, kctx, ctx);", accessor)
-                : indentedLine("MessageWire.fromWire(msgFactory, %s, kctx, ctx, clsLdr);", accessor));
-        }
-        else {
-            code.add(mode == Direction.OUT
-                ? indentedLine("MessageWire.toWire(%s, kctx, ctx);", accessor)
-                : indentedLine("MessageWire.fromWire(%s, kctx, ctx, clsLdr);", accessor));
-        }
-
-        indent--;
-
-        return code;
-    }
-
-    /** Generates a null-and-ctx-guarded {@code marshal/unmarshal} call on a {@code CacheObject} (marshal or cache-aware unmarshal only). */
-    private List<String> cacheObjectCode(String accessor, Direction mode) {
-        List<String> code = new ArrayList<>();
-
-        code.add(indentedLine("if (%s != null && ctx != null)", accessor));
-
-        indent++;
-
-        code.add(mode == Direction.OUT
-            ? indentedLine("%s.marshal(ctx);", accessor)
-            : indentedLine("%s.unmarshal(ctx, clsLdr);", accessor));
-
-        indent--;
-
-        return code;
-    }
-
-    /** Generates a null-guarded for-each loop over the array's elements. */
-    private List<String> arrayCode(TypeMirror comp, String accessor, Direction mode) {
-        Element elem = ((DeclaredType)comp).asElement();
-
-        indent++;
-
-        List<String> loopCode = forLoop(elem.getSimpleName().toString(), comp, accessor, mode);
-
-        indent--;
-
-        if (!loopCode.isEmpty())
-            imports.add(((QualifiedNameable)elem).getQualifiedName().toString());
-
-        return wrapNullGuarded(accessor, loopCode);
-    }
-
-    /** Generates a null-guarded for-each loop over the collection's elements. */
-    private List<String> collectionCode(DeclaredType t, String accessor, Direction mode) {
-        TypeMirror arg = t.getTypeArguments().get(0);
-
-        if (arg.getKind() != TypeKind.DECLARED && arg.getKind() != TypeKind.TYPEVAR)
-            return List.of();
-
-        Element elem = element(arg);
-
-        String typeName = elem.getSimpleName().toString();
-
-        indent++;
-
-        List<String> loopCode = forLoop(typeName, arg, "(Collection<? extends " + typeName + ">)" + accessor, mode);
-
-        indent--;
-
-        if (!loopCode.isEmpty()) {
-            imports.add(((QualifiedNameable)elem).getQualifiedName().toString());
-            imports.add(Collection.class.getName());
-        }
-
-        return wrapNullGuarded(accessor, loopCode);
-    }
-
-    /** Iterates {@code keySet()} then {@code values()}, wrapping both loops in a null-guard. */
-    private List<String> mapCode(DeclaredType t, String accessor, Direction mode) {
-        List<? extends TypeMirror> args = t.getTypeArguments();
-
-        indent++;
-
-        List<String> combined = new ArrayList<>();
-
-        for (int i = 0; i < 2; i++) {
-            TypeMirror elemType = args.get(i);
-
-            if (elemType.getKind() != TypeKind.DECLARED && elemType.getKind() != TypeKind.TYPEVAR)
-                continue;
-
-            Element elem = element(elemType);
-
-            String typeName = elem.getSimpleName().toString();
-            String collection = i == 0 ? "keySet" : "values";
-            String iterable = "((Collection<? extends " + typeName + ">)" + accessor + "." + collection + "())";
-
-            List<String> loopCode = forLoop(typeName, elemType, iterable, mode);
-
-            if (loopCode.isEmpty())
-                continue;
-
-            imports.add(((QualifiedNameable)elem).getQualifiedName().toString());
-            imports.add(Collection.class.getName());
-
-            combined.addAll(loopCode);
-        }
-
-        indent--;
-
-        return wrapNullGuarded(accessor, combined);
-    }
-
-    /** Returns empty if {@code elemType} requires no marshalling; otherwise returns a for-each loop over {@code iterable}. */
-    private List<String> forLoop(String typeName, TypeMirror elemType, String iterable, Direction mode) {
-        String el = loopDepth == 0 ? "e" : "e" + loopDepth;
-
-        loopDepth++;
-        indent++;
-
-        List<String> inner = codeFor(elemType, el, mode);
-
-        indent--;
-        loopDepth--;
-
-        if (inner.isEmpty())
-            return List.of();
-
-        List<String> code = new ArrayList<>();
-
-        code.add(indentedLine("for (%s %s : %s) {", typeName, el, iterable));
-
-        code.addAll(inner);
-
-        code.add(indentedLine("}"));
-
-        return code;
-    }
-
-    /** Prefixes {@code body} with the {@code msgFactory} resolution line when a loop-nested facade call was emitted. */
-    private void prependMsgFactoryResolution(List<String> body) {
-        if (!usesMsgFactory)
-            return;
-
-        imports.add(IGNITE_MESSAGE_FACTORY_CLS);
-
-        body.add(0, EMPTY);
-        body.add(0, indentedLine("IgniteMessageFactory msgFactory = (IgniteMessageFactory)kctx.messageFactory();"));
-    }
-
-    /** Returns empty if {@code inner} is empty; otherwise wraps {@code inner} in a null-guard on {@code nullGuard}. */
-    private List<String> wrapNullGuarded(String nullGuard, List<String> inner) {
-        if (inner.isEmpty())
-            return List.of();
-
-        List<String> code = new ArrayList<>();
-
-        code.add(indentedLine("if (%s != null) {", nullGuard));
-
-        code.addAll(inner);
-
-        code.add(indentedLine("}"));
-
-        return code;
-    }
-
-    /**
-     * Returns the {@code CacheObjectContext ctx} resolution line for the current message type. Cache messages resolve
-     * via the cache, group messages via the cache group — the group's context outlives the stop of individual caches,
-     * so cache objects still unmarshal while a cache (group) is being destroyed.
-     */
-    private String ctxResolutionLine() {
-        if (isCacheIdAwareMessage(type))
-            return indentedLine("CacheObjectContext ctx = cacheObjCtx != null ? cacheObjCtx : " +
-                    "kctx.cache().context().cacheObjectContext(msg.cacheId());");
-        else if (isCacheGroupIdMessage(type))
-            return indentedLine("CacheObjectContext ctx = cacheObjCtx != null ? cacheObjCtx : " +
-                    "kctx.cache().cacheGroup(msg.groupId()) == null ? null : " +
-                    "kctx.cache().cacheGroup(msg.groupId()).cacheObjectContext();");
-        else
-            return indentedLine("CacheObjectContext ctx = cacheObjCtx;");
-    }
-
-    /** Returns {@code true} if any field requires {@code ctx} in generated marshal/unmarshal code. */
-    private boolean needsCtx(List<VariableElement> fields) {
-        return fields.stream().anyMatch(f -> needsCtxType(f.asType()));
-    }
-
-    /**
-     * Returns whether the {@code @Order} fields of {@code msgType} need a cache object context to unmarshal. Such a
-     * message must not be a {@code @NioField}: its {@code unmarshalNio} runs on the NIO thread, which has no context.
-     * A type with no fields to inspect (e.g. a type variable) is conservatively assumed to need the context.
-     */
-    private boolean nestedNeedsCtx(TypeMirror type) {
-        Element el = env.getTypeUtils().asElement(type);
-
-        if (!(el instanceof TypeElement))
-            return true;
-
-        return SystemViewRowAttributeWalkerProcessor.superclasses(env, (TypeElement)el)
-            .flatMap(c -> ElementFilter.fieldsIn(c.getEnclosedElements()).stream())
-            .filter(f -> f.getAnnotation(Order.class) != null)
-            .anyMatch(f -> needsCtxType(f.asType()));
-    }
-
-    /** Returns {@code true} if type {@code t} (or its element/key/value types) requires {@code ctx}. */
-    private boolean needsCtxType(TypeMirror t) {
-        if (t.getKind() == TypeKind.ARRAY)
-            return needsCtxType(((ArrayType)t).getComponentType());
-
-        if (t.getKind() == TypeKind.DECLARED || t.getKind() == TypeKind.TYPEVAR) {
-            if (isMessage(t))
-                return !isNonMarshallable(t);
-
-            if (isCacheObject(t))
-                return true;
-
-            if (isMap(t)) {
-                List<? extends TypeMirror> args = ((DeclaredType)t).getTypeArguments();
-                return needsCtxType(args.get(0)) || needsCtxType(args.get(1));
-            }
-
-            if (isCollection(t)) {
-                List<? extends TypeMirror> args = ((DeclaredType)t).getTypeArguments();
-                return needsCtxType(args.get(0));
-            }
-        }
-
-        return false;
-    }
-
-    /** */
-    private boolean isMessage(TypeMirror type) {
-        return assignableFrom(type, msgType);
-    }
-
-    /** */
-    private boolean isCacheObject(TypeMirror type) {
-        return assignableFrom(type, cacheObjType);
-    }
-
-    /** Returns {@code true} if {@code type} (erased) is assignable to {@code java.util.Map}. */
-    private boolean isMap(TypeMirror type) {
-        return assignableFrom(erasedType(type), mapType);
-    }
-
-    /** Returns {@code true} if {@code type} (erased) is assignable to {@code java.util.Collection}. */
-    private boolean isCollection(TypeMirror type) {
-        return assignableFrom(erasedType(type), colType);
-    }
-
-    /** Recursion skip for such fields is subtype-safe: subclasses inherit the {@code NonMarshallableMessage} marker. */
-    private boolean isNonMarshallable(TypeMirror t) {
-        return assignableFrom(t, nonMarshallableType);
-    }
-
-    /** */
-    private boolean isCacheGroupIdMessage(TypeElement te) {
-        return assignableFrom(te.asType(), cacheGrpIdMsgType);
-    }
-
-    /** */
-    private static boolean isNioField(VariableElement field) {
-        return field.getAnnotation(NioField.class) != null;
-    }
-
-    /** Marshalling flavour of a {@code @Marshalled} field, told apart by the shape of its companion wire field(s). */
-    private enum MarshalledKind {
-        /** {@code byte[]} companion: the whole object is a single marshaller blob. */
-        BLOB,
-
-        /** {@code Message[]} companion: per-element {@code Message} serialization, the collection is rebuilt on unmarshal. */
-        ELEMENTS,
-
-        /** {@code Collection<byte[]>} companion: per-element marshaller blobs, each element keeping its own class loader. */
-        ELEMENT_BLOBS,
-
-        /** Two companions ({@code keys()}/{@code values()}): a {@code Map} serialized as parallel wire fields. */
-        MAP
-    }
-
-    /**
-     * @return the flavour of {@code field}'s {@code @Marshalled}, or {@code null} when the field is not annotated.
-     * Called once per field when building {@link #kinds}; look the kind up there instead.
-     */
-    private @Nullable MarshalledKind marshalledKind(VariableElement field) {
-        Marshalled ann = field.getAnnotation(Marshalled.class);
-
-        if (ann == null)
-            return null;
-
-        boolean map = !ann.keys().isEmpty() || !ann.values().isEmpty();
-
-        if (map == !ann.value().isEmpty() || (map && (ann.keys().isEmpty() || ann.values().isEmpty()))) {
-            env.getMessager().printMessage(Diagnostic.Kind.ERROR,
-                "@Marshalled must set either value() or both keys() and values()", field);
-
-            return null;
-        }
-
-        if (map)
-            return MarshalledKind.MAP;
-
-        TypeMirror wire = requireEnclosed(enclosed, ann.value(), "@Marshalled").asType();
-
-        if (wire.getKind() == TypeKind.ARRAY) {
-            return ((ArrayType)wire).getComponentType().getKind() == TypeKind.BYTE
-                ? MarshalledKind.BLOB
-                : MarshalledKind.ELEMENTS;
-        }
-
-        return MarshalledKind.ELEMENT_BLOBS;
-    }
-
-    /** Returns the enclosed field named {@code name}, or throws if absent. */
-    private VariableElement requireEnclosed(Map<String, VariableElement> enclosed, String name, String annotationName) {
-        VariableElement el = enclosed.get(name);
-
-        if (el == null)
-            throw new IllegalStateException(annotationName + " companion field '" + name + "' not found in " + type);
-
-        return el;
-    }
-
-    /** Iterates all {@code @Marshalled} fields and applies {@code codeGen(bytesAccessor, objAccessor)} to each. */
-    private void forEachMarshalled(BiFunction<String, String, List<String>> codeGen, List<String> body) {
-        for (VariableElement field : enclosed.values()) {
-            if (kinds.get(field) != MarshalledKind.BLOB)
-                continue;
-
-            Marshalled ann = field.getAnnotation(Marshalled.class);
-
-            appendBlock(body, codeGen.apply("msg." + ann.value(), "msg." + field.getSimpleName()));
-        }
-    }
-
-    /** Returns the element for {@code t}; for a type variable, uses its upper bound. */
-    private Element element(TypeMirror t) {
-        return t.getKind() == TypeKind.DECLARED
-            ? ((DeclaredType)t).asElement()
-            : ((DeclaredType)((TypeVariable)t).getUpperBound()).asElement();
-    }
-
-    /** Returns the simple name of the array component type of {@code field}, registering its import. */
-    private String arrayComponentName(VariableElement field) {
-        Element comp = ((DeclaredType)((ArrayType)field.asType()).getComponentType()).asElement();
-
-        imports.add(((QualifiedNameable)comp).getQualifiedName().toString());
-
-        return comp.getSimpleName().toString();
-    }
-
-    /** Direction of the field code a generator pass emits: object→wire ({@link #MARSHAL}) or wire→object ({@link #UNMARSHAL}). */
-    private enum Direction {
-        /** On the way out: the code runs before the message is written. */
-        OUT,
-
-        /** On the way in: the code runs after the message is read, with cache context and class loader at hand. */
-        IN
     }
 }
