+++ modules/core/src/main/java/org/apache/ignite/internal/managers/communication/MessageWire.java
@@ -22,18 +22,21 @@ import org.apache.ignite.internal.CustomWireFormMessage;
 import org.apache.ignite.internal.GridKernalContext;
 import org.apache.ignite.internal.processors.cache.CacheObjectContext;
 import org.apache.ignite.plugin.extensions.communication.Message;
+import org.apache.ignite.plugin.extensions.communication.MessageMarshaller;
 import org.apache.ignite.plugin.extensions.communication.MessageSerializer;
+import org.apache.ignite.plugin.extensions.communication.MessageWireForm;
 import org.jetbrains.annotations.Nullable;
 
 /**
  * Takes a {@link Message} to the state it is written from, and back to the state it is used in. Two steps take part,
  * and this class is the only place that knows both and their order:
  * <ul>
- *     <li>marshalling, done by the generated marshaller — see {@code MessageMarshalling};</li>
- *     <li>the step a {@link CustomWireFormMessage} does itself, which is not marshalling.</li>
+ *     <li>the step a {@link CustomWireFormMessage} does itself;</li>
+ *     <li>{@link MessageMarshaller} — the fields that become bytes;</li>
+ *     <li>{@link MessageWireForm} — walking the nested messages and cache objects.</li>
  * </ul>
- * The message's own step runs first on the way out and last on the way in, so marshalling always deals with the
- * fields as they go on the wire. Writing and reading the bytes themselves is left to {@link MessageSerializer}.
+ * On the way out they run in that order, on the way in in reverse, so each step sees the fields in the shape the
+ * next one expects. Writing and reading the bytes themselves is left to {@link MessageSerializer}.
  */
 public final class MessageWire {
     /** */
@@ -50,25 +53,31 @@ public final class MessageWire {
      */
     public static <M extends Message> void toWire(M msg, GridKernalContext kctx, @Nullable CacheObjectContext cacheObjCtx)
         throws IgniteCheckedException {
-        wireForm(msg, true);
-
-        MessageMarshalling.marshal(msg, kctx, cacheObjCtx);
+        toWire(factory(kctx), msg, kctx, cacheObjCtx);
     }
 
     /**
      * Takes {@code msg} to the state it is written from. Callers doing this for many messages resolve
      * {@code msgFactory} once and use this overload.
      *
-     * @param msgFactory Message factory to resolve the marshaller from.
+     * @param msgFactory Message factory to resolve the companions from.
      * @param msg Message to take to the wire.
      * @param kctx Kernal context.
      * @param cacheObjCtx Cache object context of the enclosing message, or {@code null} at the top level.
      */
     public static <M extends Message> void toWire(IgniteMessageFactory msgFactory, M msg, GridKernalContext kctx,
         @Nullable CacheObjectContext cacheObjCtx) throws IgniteCheckedException {
-        wireForm(msg, true);
+        ownStep(msg, true);
 
-        MessageMarshalling.marshal(msgFactory, msg, kctx, cacheObjCtx);
+        MessageMarshaller<M> marshaller = marshaller(msgFactory, msg);
+
+        if (marshaller != null)
+            marshaller.marshal(msg, kctx, cacheObjCtx);
+
+        MessageWireForm<M> wireForm = wireForm(msgFactory, msg);
+
+        if (wireForm != null)
+            wireForm.toWire(msg, kctx, cacheObjCtx);
     }
 
     /**
@@ -81,16 +90,14 @@ public final class MessageWire {
      */
     public static <M extends Message> void fromWire(M msg, GridKernalContext kctx, @Nullable CacheObjectContext cacheObjCtx,
         ClassLoader clsLdr) throws IgniteCheckedException {
-        MessageMarshalling.unmarshal(msg, kctx, cacheObjCtx, clsLdr);
-
-        wireForm(msg, false);
+        fromWire(factory(kctx), msg, kctx, cacheObjCtx, clsLdr);
     }
 
     /**
      * Brings {@code msg} back to the state it is used in. Callers doing this for many messages resolve
      * {@code msgFactory} once and use this overload.
      *
-     * @param msgFactory Message factory to resolve the marshaller from.
+     * @param msgFactory Message factory to resolve the companions from.
      * @param msg Message to bring back.
      * @param kctx Kernal context.
      * @param cacheObjCtx Cache object context of the enclosing message, or {@code null} at the top level.
@@ -98,9 +105,20 @@ public final class MessageWire {
      */
     public static <M extends Message> void fromWire(IgniteMessageFactory msgFactory, M msg, GridKernalContext kctx,
         @Nullable CacheObjectContext cacheObjCtx, ClassLoader clsLdr) throws IgniteCheckedException {
-        MessageMarshalling.unmarshal(msgFactory, msg, kctx, cacheObjCtx, clsLdr);
+        assert !MessageUnmarshalOnceCheck.ENABLED || MessageUnmarshalOnceCheck.firstUnmarshal(msg, true)
+            : "Finish-unmarshalled more than once: " + msg.getClass().getName();
 
-        wireForm(msg, false);
+        MessageWireForm<M> wireForm = wireForm(msgFactory, msg);
+
+        if (wireForm != null)
+            wireForm.fromWire(msg, kctx, cacheObjCtx, clsLdr);
+
+        MessageMarshaller<M> marshaller = marshaller(msgFactory, msg);
+
+        if (marshaller != null)
+            marshaller.unmarshal(msg, kctx, cacheObjCtx, clsLdr);
+
+        ownStep(msg, false);
     }
 
     /**
@@ -111,20 +129,36 @@ public final class MessageWire {
      * @param kctx Kernal context.
      */
     public static <M extends Message> void fromWire(M msg, GridKernalContext kctx) throws IgniteCheckedException {
-        MessageMarshalling.unmarshal(msg, kctx);
+        assert !MessageUnmarshalOnceCheck.ENABLED || MessageUnmarshalOnceCheck.firstUnmarshal(msg, false)
+            : "Finish-unmarshalled more than once: " + msg.getClass().getName();
 
-        wireForm(msg, false);
+        IgniteMessageFactory msgFactory = factory(kctx);
+
+        MessageWireForm<M> wireForm = wireForm(msgFactory, msg);
+
+        if (wireForm != null)
+            wireForm.fromWire(msg, kctx);
+
+        MessageMarshaller<M> marshaller = marshaller(msgFactory, msg);
+
+        if (marshaller != null)
+            marshaller.unmarshal(msg, kctx);
+
+        ownStep(msg, false);
     }
 
     /**
      * Brings back only the {@code @NioField} fields (routing headers) on the NIO thread. The rest of the message,
-     * the message's own step included, waits for a worker thread.
+     * the other steps included, waits for a worker thread.
      *
      * @param msg Message to bring back.
      * @param kctx Kernal context.
      */
     public static <M extends Message> void fromWireNio(M msg, GridKernalContext kctx) throws IgniteCheckedException {
-        MessageMarshalling.unmarshalNio(msg, kctx);
+        MessageWireForm<M> wireForm = wireForm(factory(kctx), msg);
+
+        if (wireForm != null)
+            wireForm.fromWireNio(msg, kctx);
     }
 
     /**
@@ -133,7 +167,7 @@ public final class MessageWire {
      * @param msg Message to run the step of.
      * @param out {@code true} on the way out, {@code false} on the way in.
      */
-    private static void wireForm(Message msg, boolean out) {
+    private static void ownStep(Message msg, boolean out) {
         if (!(msg instanceof CustomWireFormMessage))
             return;
 
@@ -141,5 +175,22 @@ public final class MessageWire {
             ((CustomWireFormMessage)msg).toWireForm();
         else
             ((CustomWireFormMessage)msg).fromWireForm();
+    }
+
+    /** @return the message factory of {@code kctx}. */
+    private static IgniteMessageFactory factory(GridKernalContext kctx) {
+        return (IgniteMessageFactory)kctx.messageFactory();
+    }
+
+    /** @return the marshaller registered for {@code msg}'s direct type, or {@code null} if none. */
+    @SuppressWarnings("unchecked")
+    private static <M extends Message> @Nullable MessageMarshaller<M> marshaller(IgniteMessageFactory msgFactory, M msg) {
+        return (MessageMarshaller<M>)msgFactory.marshaller(msg.directType());
+    }
+
+    /** @return the wire form registered for {@code msg}'s direct type, or {@code null} if none. */
+    @SuppressWarnings("unchecked")
+    private static <M extends Message> @Nullable MessageWireForm<M> wireForm(IgniteMessageFactory msgFactory, M msg) {
+        return (MessageWireForm<M>)msgFactory.wireForm(msg.directType());
     }
 }
