+++ modules/core/src/main/java/org/apache/ignite/spi/IgniteSpiAdapter.java
@@ -54,6 +54,7 @@ import org.apache.ignite.plugin.extensions.communication.MessageFormatter;
 import org.apache.ignite.plugin.extensions.communication.MessageMarshaller;
 import org.apache.ignite.plugin.extensions.communication.MessageReader;
 import org.apache.ignite.plugin.extensions.communication.MessageSerializer;
+import org.apache.ignite.plugin.extensions.communication.MessageWireForm;
 import org.apache.ignite.plugin.extensions.communication.MessageWriter;
 import org.apache.ignite.plugin.security.SecuritySubject;
 import org.apache.ignite.resources.IgniteInstanceResource;
@@ -762,6 +763,7 @@ public abstract class IgniteSpiAdapter implements IgniteSpi {
                         short directType,
                         MessageSerializer<Message> serializer,
                         @Nullable MessageMarshaller<Message> marshaller,
+                        @Nullable MessageWireForm<Message> wireForm,
                         @Nullable GridCacheMessageDeployer<GridCacheMessage> deployer
                     ) throws IgniteException {
                         throw new IgniteException("Failed to register message, node is not started.");
@@ -776,6 +778,10 @@ public abstract class IgniteSpiAdapter implements IgniteSpi {
                     }
 
                     @Nullable @Override public MessageMarshaller<Message> marshaller(short type) {
+                        throw new IgniteException("Failed to register message, node is not started.");
+                    }
+
+                    @Nullable @Override public MessageWireForm<Message> wireForm(short type) {
                         throw new IgniteException("Failed to register message, node is not started.");
                     }
