+++ modules/core/src/test/resources/codegen/TestCustomWireFormMessageSerializer.java
