+++ modules/core/src/test/resources/codegen/TestMarshallableMessageSerializer.java
