+++ modules/core/src/test/resources/codegen/TestMarshalledCollectionMessageMarshaller.java
@@ -40,7 +40,6 @@ public final class TestMarshalledCollectionMessageMarshaller implements MessageM
     /** */
     @Override public void unmarshal(TestMarshalledCollectionMessage msg, GridKernalContext kctx, CacheObjectContext cacheObjCtx, ClassLoader clsLdr) throws IgniteCheckedException {
         CacheObjectContext ctx = cacheObjCtx;
-
         if (msg.keysArr != null) {
             msg.keys = U.newHashSet(msg.keysArr.length);
