+++ modules/core/src/test/resources/codegen/TestMarshalledObjectsMessageMarshaller.java
@@ -55,7 +55,6 @@ public final class TestMarshalledObjectsMessageMarshaller implements MessageMars
     /** */
     @Override public void unmarshal(TestMarshalledObjectsMessage msg, GridKernalContext kctx, CacheObjectContext cacheObjCtx, ClassLoader clsLdr) throws IgniteCheckedException {
         CacheObjectContext ctx = cacheObjCtx;
-
         if (msg.dataBytes != null) {
             msg.data = new ArrayList<>(msg.dataBytes.size());
